+++ packages/rip-and-tear-checkbox/checkbox.stories.ts
@@ -12,3 +12,20 @@ export default {
 } satisfies Meta;
 
 export const Default: StoryObj = {};
+
+export const Invalid: StoryObj = {
+  render: () => html`<w-checkbox name="invalidfoo" value="bar" invalid>Invalid checkbox</w-checkbox>`,
+};
+
+export const Disabled: StoryObj = {
+  render: () => html`<w-checkbox name="disabledfoo" value="bar" disabled>Disabled checkbox</w-checkbox>`,
+};
+
+export const BunchOfCheckboxes: StoryObj = {
+  render: () => html`
+    <w-checkbox-group>
+      <w-checkbox name="group" value="foo">Foo</w-checkbox>
+      <w-checkbox name="group" value="bar">Bar</w-checkbox>
+    </w-checkbox-group>
+  `,
+}