+++ packages/rip-and-tear-checkbox/react.ts
@@ -2,7 +2,8 @@ import { createComponent } from '@lit/react';
 import { LitElement } from 'lit';
 import React from 'react';
 
-import { WCheckbox } from './checkbox';
+import { WCheckbox } from './checkbox.js';
+import { WCheckboxGroup } from './checkbox-group.js';
 
 // decouple from CDN by providing a dummy class
 class Component extends LitElement {}
@@ -15,4 +16,10 @@ export const Checkbox = createComponent({
     onChange: 'change',
     onchange: 'change',
   },
+});
+
+export const CheckboxGroup = createComponent({
+  tagName: 'w-checkbox-group',
+  elementClass: Component as unknown as typeof WCheckboxGroup,
+  react: React,
 });