+++ packages/select/react.ts
@@ -2,7 +2,7 @@ import { LitElement } from "lit";
 import { createComponent } from "@lit/react";
 import React from "react";
 
-import { WSelect } from ".";
+import { WarpSelect } from './index.js';
 
 // decouple from CDN by providing a dummy class
 class Component extends LitElement { }